Cleanup, Fixes, and restart of codegen.
There's a bunch of cleanup for parsing here. Including fixing a couple
of bugs where we try too eagerly to form a return type for parsing,
causing entirely unrelated code to fail to parse.

Also, changes to the resultNameDecl representation are coming.
The intent of these changes is to create a "canonical" decl which
can be used by _all post conditions_ on the function to use as a
key, allowing codegen & constant evaluation to all reference the
same lvalue.